+++ Adapter/PlentymarketsAdapter/RequestGenerator/Order/Address/AddressRequestGenerator.php
@@ -62,6 +62,7 @@ class AddressRequestGenerator implements AddressRequestGeneratorInterface
             'name1' => trim($address->getCompany() . ' ' . $address->getDepartment()),
             'name2' => $address->getFirstname(),
             'name3' => $address->getLastname(),
+            'gender' => $address->getGender(),
             'postalCode' => $address->getPostalCode(),
             'town' => $address->getCity(),
             'countryId' => $countryIdentity->getAdapterIdentifier(),