+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
+### Added
+
+### Fixed
+- fix salutation in order an customer
+
 ### Changed
 - corrected the name of the isMappedIdentity function of the IdentityService